Fixes TestMetrics filter bug.
Previously, this matches implementation was ignoring the `name` passed
in and instead comparing the *literal* `"testMetricWait.pool.Wait"` to
the dynamically-created
`MetricRegistry.name("testMetricWait", "pool", "Wait")`.

This commit updates the matches implementation to instead compare the
`name` passed in.

This brings it in line with the other implementations in the same file.